refactor: split indexer.rs and shelves/read.rs under the 800-line cap (#1452)
Both files exceeded the ~800-line soft cap in .claude/rules/05-rust-style.md.
Pure module reorganization, no schema or behavior change:

- db/src/indexer.rs (847 lines) -> indexer/{mod,audiobooks,backfill}.rs:
  mod.rs keeps the ebook reindex entry point + diff_library; audiobooks.rs
  holds the audiobook-reindex pipeline; backfill.rs holds the chapter and
  word-count backfill passes.
- db/src/shelves/read.rs (837 lines) -> read/{mod,summary,detail}.rs, split
  by read-shape: mod.rs holds shared constants/helpers, summary.rs is the
  multi-shelf rail (list_visible_shelves + batch-loaders), detail.rs is
  single-shelf reads (get_shelf, shelf_page, preview_rule, Kobo sync
  membership).

docs/architecture.md's db module map updated for both new submodule layouts.
