+++ db/src/indexer/mod.rs
@@ -8,7 +8,13 @@ use std::path::{Path, PathBuf};
 
 use sqlx::SqlitePool;
 
-use crate::{audiobook, books, ebook, sync};
+use crate::{books, ebook, sync};
+
+mod audiobooks;
+mod backfill;
+
+pub use audiobooks::{reindex_audiobooks, reindex_audiobooks_with_progress};
+pub(crate) use backfill::{backfill_chapters, backfill_word_counts};
 
 /// Errors returned by the pure-DB indexer reads (`is_stale`). The
 /// transparent `Db` variant honors the `02-error-handling` boundary rule
@@ -457,390 +463,6 @@ async fn gc_missing_files_best_effort(pool: &SqlitePool) {
         Ok(_) => {}
         Err(e) => tracing::error!("reindex: missing-files GC failed: {e}"),
     }
-}
-
-/// Audiobook-library sibling of [`reindex`]. Groups audio files by folder,
-/// reads multi-part tags, then calls [`sync::sync_audiobooks`] to write
-/// `book_file_parts` rows.
-pub async fn reindex_audiobooks(
-    pool: &SqlitePool,
-    library_path: &str,
-) -> anyhow::Result<ReindexStats> {
-    reindex_audiobooks_with_progress(pool, library_path, |_, _| {}).await
-}
-
-/// Enumeration-trust signals lifted out of Phase A so the caller can gate
-/// the removal pass without re-reading the filesystem.
-struct EnumerationSignals {
-    /// A subdir `read_dir` failed — partial view.
-    incomplete: bool,
-    /// The walk saw at least one regular file (of any extension).
-    saw_any_file: bool,
-}
-
-/// Phase A of [`reindex_audiobooks_with_progress`]: stat every audio file
-/// under `library_path`, then group the per-file entries into one
-/// [`audiobook::AudiobookGroup`] per book (folder-based grouping). The
-/// [`EnumerationSignals`] ride alongside so the caller can suppress the
-/// removal pass on a partial or empty scan.
-async fn stat_and_group_audiobooks(
-    library_path: &str,
-) -> anyhow::Result<(Vec<audiobook::AudiobookGroup>, EnumerationSignals)> {
-    let path_for_scan = library_path.to_owned();
-    let library_key = library_path.to_owned();
-    let stat = tokio::task::spawn_blocking(move || {
-        audiobook::stat_audiobook_library(Some(&path_for_scan), &library_key)
-    })
-    .await?;
-    if let Some(msg) = stat.error {
-        anyhow::bail!("audiobook scan of {library_path} failed: {msg}");
-    }
-
-    let signals = EnumerationSignals {
-        incomplete: stat.incomplete,
-        saw_any_file: stat.saw_any_file,
-    };
-    let entries = stat.entries;
-    let library_key2 = library_path.to_owned();
-    let groups =
-        tokio::task::spawn_blocking(move || audiobook::group_into_books(entries, &library_key2))
-            .await?;
-    Ok((groups, signals))
-}
-
-/// Project audiobook groups to the ebook [`ebook::StatEntry`] shape so
-/// `diff_library` can be reused verbatim across both library kinds. Skips
-/// attachment-only groups (empty `scan_key`).
-fn project_groups_to_stat(groups: &[audiobook::AudiobookGroup]) -> Vec<ebook::StatEntry> {
-    groups
-        .iter()
-        .filter(|g| !g.scan_key.is_empty())
-        .map(|g| ebook::StatEntry {
-            filename: g.group_path.clone(),
-            scan_key: g.scan_key.clone(),
-            mtime_epoch: g.max_mtime_epoch,
-            size_bytes: g.total_size_bytes,
-            error: None,
-        })
-        .collect()
-}
-
-/// [`reindex_audiobooks`] variant that calls `on_progress(processed,
-/// total)` after each per-book write inside `sync_audiobooks`. Used by
-/// [`crate::worker::Worker`] to report determinate `processed / total`
-/// counts to the UI indicator. Returns the scan's ghost-count tallies
-/// (issue #1057) so the caller can decide whether to attach a warning.
-pub async fn reindex_audiobooks_with_progress(
-    pool: &SqlitePool,
-    library_path: &str,
-    on_progress: impl FnMut(u32, u32),
-) -> anyhow::Result<ReindexStats> {
-    let (groups, signals) = stat_and_group_audiobooks(library_path).await?;
-
-    // Diff groups against DB rows (project groups to the ebook StatEntry shape
-    // so diff_library can be reused verbatim). Scope to audiobook formats so a
-    // shared ebook + audiobook library_path does not classify EPUB rows here
-    // as Removed (#328).
-    let mut db_rows =
-        books::list_indexed_rows_for_formats(pool, library_path, audiobook::AUDIOBOOK_FORMATS)
-            .await?;
-    // Merged/attached audiobook files diff against their book_files stat
-    // via merged_uuids — same as the ebook path above.
-    db_rows.extend(
-        books::list_merged_rows_for_formats(pool, library_path, audiobook::AUDIOBOOK_FORMATS)
-            .await?,
-    );
-    let library_root: PathBuf = PathBuf::from(library_path);
-    let groups_as_stat = project_groups_to_stat(&groups);
-    let db_file_backed = db_rows.iter().filter(|r| r.has_file).count();
-    let trustworthy =
-        enumeration_is_trustworthy(signals.incomplete, signals.saw_any_file, db_file_backed > 0);
-    if !trustworthy {
-        tracing::warn!(
-            library_path,
-            incomplete = signals.incomplete,
-            saw_any_file = signals.saw_any_file,
-            db_file_backed,
-            "reindex_audiobooks: enumeration incomplete or a populated root read empty — \
-             skipping the removal pass; no books marked missing (issue #819)"
-        );
-    }
-    let diff = diff_library(&groups_as_stat, &db_rows, &library_root, trustworthy);
-    let removed_count = diff.removed.len();
-    check_mass_missing(removed_count, db_file_backed)?;
-
-    // Phase B: parse only the New and Changed groups.
-    let groups_by_group_path: std::collections::HashMap<String, audiobook::AudiobookGroup> = groups
-        .into_iter()
-        .filter(|g| !g.scan_key.is_empty())
-        .map(|g| (g.group_path.clone(), g))
-        .collect();
-
-    let new_groups: Vec<audiobook::AudiobookGroup> = diff
-        .new
-        .iter()
-        .filter_map(|t| groups_by_group_path.get(&t.filename).cloned())
-        .collect();
-    let changed_groups: Vec<audiobook::AudiobookGroup> = diff
-        .changed
-        .iter()
-        .filter_map(|t| groups_by_group_path.get(&t.filename).cloned())
-        .collect();
-
-    let root_for_parse = library_root.clone();
-    let parsed = tokio::task::spawn_blocking(move || {
-        let new_books = audiobook::parse_groups(new_groups, &root_for_parse);
-        let changed_books = audiobook::parse_groups(changed_groups, &root_for_parse);
-        (new_books, changed_books)
-    })
-    .await?;
-
-    let plan = sync::AudiobookSyncPlan {
-        new_books: parsed.0,
-        changed_books: parsed.1,
-        removed_uuids: diff.removed,
-        backfill: diff.backfill,
-    };
-    sync::sync_audiobooks_with_progress(pool, library_path, plan, on_progress).await?;
-    gc_missing_files_best_effort(pool).await;
-    Ok(ReindexStats {
-        removed: removed_count,
-        file_backed_total: db_file_backed,
-    })
-}
-
-/// Query the first-part filename (ordinal=0) and format for every book
-/// under `library_path` that needs chapter backfill (no `file_chapters`
-/// rows yet).
-async fn fetch_backfill_candidates(
-    pool: &SqlitePool,
-    library_path: &str,
-) -> anyhow::Result<Vec<(i64, String, String)>> {
-    let rows: Vec<(i64, String, String)> = sqlx::query_as(
-        "SELECT bf.id, bfp.filename, bf.format \
-         FROM book_files bf \
-         JOIN books b ON bf.book_id = b.id \
-         JOIN scan_roots l ON b.library_id = l.id \
-         JOIN book_file_parts bfp ON bfp.book_file_id = bf.id \
-         WHERE l.path = ? \
-           AND bf.format IN ('M4B', 'M4A', 'MP3') \
-           AND bfp.ordinal = 0 \
-           AND NOT EXISTS (SELECT 1 FROM file_chapters fc WHERE fc.book_file_id = bf.id) \
-         ORDER BY bf.id",
-    )
-    .bind(library_path)
-    .fetch_all(pool)
-    .await?;
-    Ok(rows)
-}
-
-/// Bulk-fetch all `book_file_parts` rows for `book_file_ids` and group them
-/// by `book_file_id` — avoids N per-book SELECT round-trips. Chunked at 500
-/// to stay well under SQLite's 999 bind-parameter limit.
-async fn bulk_fetch_parts(
-    pool: &SqlitePool,
-    book_file_ids: &[i64],
-) -> anyhow::Result<std::collections::HashMap<i64, Vec<crate::audiobook::AudiobookPart>>> {
-    let mut all_parts_rows: Vec<(i64, i64, String, i64, i64, f64)> = Vec::new();
-    for chunk in book_file_ids.chunks(500) {
-        let placeholders = std::iter::repeat_n("?", chunk.len())
-            .collect::<Vec<_>>()
-            .join(", ");
-        let parts_sql = format!(
-            "SELECT book_file_id, ordinal, filename, size_bytes, mtime_epoch, duration_seconds \
-             FROM book_file_parts WHERE book_file_id IN ({placeholders}) \
-             ORDER BY book_file_id, ordinal"
-        );
-        let mut parts_query = sqlx::query_as::<_, (i64, i64, String, i64, i64, f64)>(&parts_sql);
-        for id in chunk {
-            parts_query = parts_query.bind(id);
-        }
-        all_parts_rows.extend(parts_query.fetch_all(pool).await?);
-    }
-
-    let mut parts_by_id: std::collections::HashMap<i64, Vec<crate::audiobook::AudiobookPart>> =
-        std::collections::HashMap::new();
-    for (book_file_id, ordinal, filename, size_bytes, mtime_epoch, duration_seconds) in
-        all_parts_rows
-    {
-        parts_by_id
-            .entry(book_file_id)
-            .or_default()
-            .push(crate::audiobook::AudiobookPart {
-                ordinal,
-                filename,
-                size_bytes,
-                mtime_epoch,
-                duration_seconds,
-            });
-    }
-    Ok(parts_by_id)
-}
-
-/// Fill `file_chapters` for audiobook `book_files` rows that have none.
-///
-/// The chapter extraction pipeline was added after the initial audiobook
-/// indexer, so books indexed before the migration have zero `file_chapters`
-/// rows. The normal diff-based reindex skips unchanged files, so this
-/// backfill runs as a separate worker task and is a no-op once all books
-/// have chapters. `on_progress(processed, total)` is called after each
-/// book so the UI can render a progress bar.
-///
-/// ## Query efficiency
-///
-/// All `book_file_parts` rows for the backfill set are fetched in a single
-/// `WHERE book_file_id IN (…)` bulk query before the loop rather than one
-/// per book, and all chapter inserts are committed in batches of 250 books
-/// to avoid per-book WAL flushes (mirrors the sync/audiobooks.rs backfill
-/// pattern).
-pub(crate) async fn backfill_chapters(
-    pool: &SqlitePool,
-    library_path: &str,
-    mut on_progress: impl FnMut(u32, u32),
-) -> anyhow::Result<()> {
-    let rows = fetch_backfill_candidates(pool, library_path).await?;
-    if rows.is_empty() {
-        return Ok(());
-    }
-
-    let total = u32::try_from(rows.len()).unwrap_or(u32::MAX);
-    tracing::info!(
-        count = total,
-        "backfilling chapters for existing audiobooks"
-    );
-
-    let book_file_ids: Vec<i64> = rows.iter().map(|(id, _, _)| *id).collect();
-    let parts_by_id = bulk_fetch_parts(pool, &book_file_ids).await?;
-
-    // Process books in batches of 250 to bound transaction size (mirrors the
-    // sync/audiobooks.rs backfill pattern).
-    let lib_root = PathBuf::from(library_path);
-    for (batch_idx, chunk) in rows.chunks(250).enumerate() {
-        let mut tx = pool.begin().await?;
-        for (i, (book_file_id, first_part_filename, format)) in chunk.iter().enumerate() {
-            let abs = lib_root.join(first_part_filename);
-            let fmt = format.clone();
-            let chapters =
-                tokio::task::spawn_blocking(move || audiobook::extract_chapters(&abs, &fmt))
-                    .await
-                    .unwrap_or_else(|join_err| {
-                        tracing::warn!(
-                            book_file_id,
-                            %join_err,
-                            is_panic = join_err.is_panic(),
-                            "chapter extraction task failed; using synthetic fallback"
-                        );
-                        Vec::new()
-                    });
-
-            let parts = parts_by_id
-                .get(book_file_id)
-                .map(Vec::as_slice)
-                .unwrap_or(&[]);
-            sync::insert_chapters(&mut tx, *book_file_id, &chapters, parts).await?;
-
-            let global_idx = batch_idx * 250 + i;
-            let progress = u32::try_from(global_idx)
-                .unwrap_or(u32::MAX)
-                .saturating_add(1);
-            on_progress(progress, total);
-        }
-        tx.commit().await?;
-    }
-
-    Ok(())
-}
-
-/// Fill `books.word_count` for EPUB-backed books under `library_path` that
-/// have none yet (NULL `word_count`).
-///
-/// Word counts were added after the initial ebook indexer, so books indexed
-/// before migration `0049` carry a NULL. The normal diff-based reindex only
-/// re-parses new/changed files, so it never revisits them; this backfill runs
-/// as a separate worker task posted after each library scan (mirroring
-/// [`backfill_chapters`]) and is a no-op once every EPUB book has a count.
-/// `on_progress(processed, total)` is called per book for the UI.
-///
-/// Each candidate's EPUB is opened and word-counted on the blocking pool
-/// (zip/XML work), and updates commit in batches of 250 to bound WAL flushes.
-/// A book whose spine can't be estimated stays NULL and is retried on the next
-/// scan — rare, and cheaper than persisting a sentinel that lies about the
-/// estimate.
-pub(crate) async fn backfill_word_counts(
-    pool: &SqlitePool,
-    library_path: &str,
-    mut on_progress: impl FnMut(u32, u32),
-) -> anyhow::Result<()> {
-    let ids = fetch_word_count_candidates(pool, library_path).await?;
-    if ids.is_empty() {
-        return Ok(());
-    }
-
-    let total = u32::try_from(ids.len()).unwrap_or(u32::MAX);
-    tracing::info!(count = total, "backfilling word counts for existing ebooks");
-
-    // One batched path lookup up front (chunked internally), same helper the
-    // stats read path used to call per-request.
-    let paths = crate::book_file_paths(pool, &ids, "EPUB").await?;
-
-    let mut processed = 0u32;
-    for chunk in ids.chunks(250) {
-        let mut tx = pool.begin().await?;
-        for &id in chunk {
-            processed = processed.saturating_add(1);
-            on_progress(processed, total);
-
-            let Some(path) = paths.get(&id).cloned() else {
-                continue;
-            };
-            let words = tokio::task::spawn_blocking(move || {
-                epub::doc::EpubDoc::new(&path)
-                    .ok()
-                    .and_then(|mut doc| ebook::estimate_word_count(&mut doc))
-            })
-            .await
-            .unwrap_or_else(|join_err| {
-                tracing::warn!(
-                    book_id = id,
-                    %join_err,
-                    is_panic = join_err.is_panic(),
-                    "word-count task failed; leaving word_count NULL"
-                );
-                None
-            });
-
-            let Some(words) = words else { continue };
-            sqlx::query("UPDATE books SET word_count = ? WHERE id = ?")
-                .bind(words)
-                .bind(id)
-                .execute(&mut *tx)
-                .await?;
-        }
-        tx.commit().await?;
-    }
-
-    Ok(())
-}
-
-/// `books.id` for every EPUB-backed book under `library_path` still missing a
-/// `word_count` — the [`backfill_word_counts`] work set. Scoped to the scanned
-/// library so the follow-up task's cost tracks that scan.
-async fn fetch_word_count_candidates(
-    pool: &SqlitePool,
-    library_path: &str,
-) -> anyhow::Result<Vec<i64>> {
-    let ids: Vec<i64> = sqlx::query_scalar(
-        "SELECT DISTINCT b.id \
-         FROM books b \
-         JOIN scan_roots l ON b.library_id = l.id \
-         JOIN book_files bf ON bf.book_id = b.id AND bf.format = 'EPUB' COLLATE NOCASE \
-         WHERE l.path = ? AND b.word_count IS NULL \
-         ORDER BY b.id",
-    )
-    .bind(library_path)
-    .fetch_all(pool)
-    .await?;
-    Ok(ids)
 }
 
 #[cfg(test)]
